poprawiona logika

+++ src/Go/Logic/Game.java
@@ -32,7 +32,7 @@ public class Game
 
 	private boolean insideBoard(int y, int x, Board board)
 	{
-		return x >= 0 && y < board.size;
+		return x >= 0 && x < board.array.length && y >= 0 && y < board.array.length;
 	}
 
 	private boolean isSurrounded(int y, int x, Board board, StoneColour colour)
@@ -41,7 +41,7 @@ public class Game
 			return true;
 		if(board.array[y][x].colour == StoneColour.Empty)
 			return false;
-		if(board.array[y][x].colour != colour)
+		if(board.array[y][x].colour != StoneColour.Empty && board.array[y][x].colour != colour)
 			return true;
 		if(board.array[y][x].visited)
 			return true;
@@ -70,6 +70,8 @@ public class Game
 			return false;
 		}
 
+		board.array[y][x].visited = false;
+
 		return true;
 	}
 
@@ -89,17 +91,89 @@ public class Game
 		kill(y, x + 1, board, colour);
 	}
 
+
+	private boolean isSelfKiller(State.Move move, Board board, StoneColour colour)
+	{
+		board.array[move.y][move.x].setColour(colour);
+
+		int y;
+		int x;
+		if(isSurrounded(move.y, move.x, board, colour))
+		{
+			y = move.y - 1;
+			x = move.x;
+			if(insideBoard(y, x, board))
+			{
+				colour = board.array[y][x].colour;
+
+				if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				{
+					board.array[move.y][move.x].setColour(StoneColour.Empty);
+					return false;
+				}
+
+			}
+			y = move.y + 1;
+			x = move.x;
+			if(insideBoard(y, x, board))
+			{
+				colour = board.array[y][x].colour;
+
+				if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				{
+					board.array[move.y][move.x].setColour(StoneColour.Empty);
+					return false;
+				}
+
+			}
+			y = move.y;
+			x = move.x - 1;
+			if(insideBoard(y, x, board))
+			{
+				colour = board.array[y][x].colour;
+
+				if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				{
+					board.array[move.y][move.x].setColour(StoneColour.Empty);
+					return false;
+				}
+
+			}
+			y = move.y;
+			x = move.x + 1;
+			if(insideBoard(y, x, board))
+			{
+				colour = board.array[y][x].colour;
+
+				if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				{
+					board.array[move.y][move.x].setColour(StoneColour.Empty);
+					return false;
+				}
+
+			}
+			board.array[move.y][move.x].setColour(StoneColour.Empty);
+			return true;
+		}
+
+		board.array[move.y][move.x].setColour(StoneColour.Empty);
+		return false;
+	}
+
+
 	public boolean isCorrect(State state, State.Move move)
 	{
 		int x = move.x;
 		int y = move.y;
-		int v = move.player.getColour().getValue();
+		StoneColour colour = move.player.getColour();
 		Board board = state.getBoard();
+
+
 
 		if(state.getPlayer() == move.player)
 			return false;
 
-		return board.array[y][x].colour.equals(StoneColour.Empty);
+		return board.array[y][x].colour.equals(StoneColour.Empty) && !isSelfKiller(move, board, colour);
 	}
 
 	private boolean enemyColour(StoneColour first, StoneColour second)
@@ -112,27 +186,44 @@ public class Game
 		int y = move.y;
 		int x = move.x;
 		Board board = state.getBoard();
-
+		StoneColour colour;
 		y = move.y - 1;
 		x = move.x;
-		StoneColour colour = board.array[y][x].colour;
-		if(isSurrounded(y, x, board, colour))
-			kill(y, x, board, colour);
+		if(insideBoard(y, x, board))
+		{
+			colour = board.array[y][x].colour;
+
+			if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				kill(y, x, board, colour);
+
+		}
 		y = move.y + 1;
 		x = move.x;
-		colour = board.array[y][x].colour;
-		if(isSurrounded(y, x, board, colour))
-			kill(y, x, board, colour);
+		if(insideBoard(y, x, board))
+		{
+			colour = board.array[y][x].colour;
+			if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				kill(y, x, board, colour);
+
+		}
 		y = move.y;
 		x = move.x - 1;
-		colour = board.array[y][x].colour;
-		if(isSurrounded(y, x, board, colour))
-			kill(y, x, board, colour);
+		if(insideBoard(y, x, board))
+		{
+			colour = board.array[y][x].colour;
+			if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				kill(y, x, board, colour);
+
+		}
 		y = move.y;
 		x = move.x + 1;
-		colour = board.array[y][x].colour;
-		if(isSurrounded(y, x, board, colour))
-			kill(y, x, board, colour);
+		if(insideBoard(y, x, board))
+		{
+			colour = board.array[y][x].colour;
+			if(colour != StoneColour.Empty && isSurrounded(y, x, board, colour))
+				kill(y, x, board, colour);
+
+		}
 
 		return new State(state.getPlayer(), board);
 	}
